should be OK

+++ src/main/java/rpn/Operator/Addition.java
@@ -6,7 +6,7 @@ public class Addition extends Operator2Operands {
     }
 
     @Override
-    public int apply(int a, int b) {
+    public double apply(double a, double b) {
         return a + b;
     }
 }
